[troff]: Trivially refactor.
* src/roff/troff/dictionary.cpp (dictionary::lookup):
* src/roff/troff/node.cpp (suppress_node::tprint): Favor function-style
  construction for temporaries over C-style type casts.

* src/roff/troff/env.cpp (environment::make_tag):
* src/roff/troff/input.cpp (set_string): Favor C++-style `static_cast`
  over C-style omnipotent casts.

* src/roff/troff/mtsm.cpp (state_set::incl, state_set::excl)
  (state_set::is_in): Discard unnecessary C-style type casts of objects
  of enumerated type to `int`.  Pre-C++11 `enum`s like these are always
  backed by `int`s.

Also stop cramming casts of return values up against the adjacent
expression.

+++ src/roff/troff/mtsm.cpp
@@ -554,62 +554,62 @@ state_set::~state_set()
 
 void state_set::incl(bool_value_state b)
 {
-  boolset |= 1 << (int)b;
+  boolset |= 1 << b;
 }
 
 void state_set::incl(int_value_state i)
 {
-  intset |= 1 << (int)i;
+  intset |= 1 << i;
 }
 
 void state_set::incl(units_value_state u)
 {
-  unitsset |= 1 << (int)u;
+  unitsset |= 1 << u;
 }
 
 void state_set::incl(string_value_state s)
 {
-  stringset |= 1 << (int)s;
+  stringset |= 1 << s;
 }
 
 void state_set::excl(bool_value_state b)
 {
-  boolset &= ~(1 << (int)b);
+  boolset &= ~(1 << b);
 }
 
 void state_set::excl(int_value_state i)
 {
-  intset &= ~(1 << (int)i);
+  intset &= ~(1 << i);
 }
 
 void state_set::excl(units_value_state u)
 {
-  unitsset &= ~(1 << (int)u);
+  unitsset &= ~(1 << u);
 }
 
 void state_set::excl(string_value_state s)
 {
-  stringset &= ~(1 << (int)s);
+  stringset &= ~(1 << s);
 }
 
 int state_set::is_in(bool_value_state b)
 {
-  return (boolset & (1 << (int)b)) != 0;
+  return (boolset & (1 << b)) != 0;
 }
 
 int state_set::is_in(int_value_state i)
 {
-  return (intset & (1 << (int)i)) != 0;
+  return (intset & (1 << i)) != 0;
 }
 
 int state_set::is_in(units_value_state u)
 {
-  return (unitsset & (1 << (int)u)) != 0;
+  return (unitsset & (1 << u)) != 0;
 }
 
 int state_set::is_in(string_value_state s)
 {
-  return (stringset & (1 << (int)s)) != 0;
+  return (stringset & (1 << s)) != 0;
 }
 
 void state_set::add(units_value_state, int n)
